changed huey to django-tasks

+++ src/newshows/helpers.py
@@ -2,6 +2,7 @@
 import logging
 import json
 import datetime
+from datetime import timedelta
 # Django second
 from django.utils.timezone import make_aware
 from django.db.models import Q
@@ -9,10 +10,8 @@ from .models import Show, ShowType, Genre, Status, Language, Country, Network, W
 # 3rd party last
 import requests
 import pendulum
-from huey import crontab
-from huey.contrib.djhuey import db_periodic_task, db_task
-
-logger = logging.getLogger(__name__)
+from loguru import logger
+from background_task import background
 
 # checked for V 0.2.0
 def _is_json(myjson):
@@ -25,8 +24,10 @@ def _is_json(myjson):
 # checked for V 0.2.0
 def _requestURL(URL, METHOD="get"):
     logger.info("Trying {}".format(URL))
-    if METHOD == "get":
+    try:
         r = requests.get(URL)
+    except:
+        return False, False
     try:
         retValue = r.json()
     except:
@@ -34,7 +35,9 @@ def _requestURL(URL, METHOD="get"):
     return r.status_code, retValue
 
 # checked for V 0.2.0
+@background(schedule=60)
 def getSonarrDownloads():
+    settings = Setting.load()
     lstDownloads = list()
     dictShow = dict()
     endpoint = "/history/"
@@ -55,11 +58,12 @@ def getSonarrDownloads():
         return False
 
 # checked for 0.2.0
-@db_periodic_task(crontab(minute='*/5'))
+@background(schedule=60)
 def checkForActiveSonarr():
     """
     checks if Sonarr is reachable and updates settings 
     """
+    settings = Setting.load()
     endpoint = "/system/status/"
     url = settings.SONARR_URL + endpoint + "?apikey=" + settings.SONARR_APIKEY
     statuscode, sonarr = _requestURL(url)
@@ -69,16 +73,17 @@ def checkForActiveSonarr():
     else:
         logger.error("Connection to Sonarr failed.")
         settings.SONARR_OK = False
+    logger.debug(f"Settings: {settings}")
     settings.save()
 
 # checked for 0.2.0
-@db_periodic_task(crontab(minute='*/5'))
+@background(schedule=300)
 def HelperUpdateSonarr():
     """
     Gets the complete list of shows in Sonarr API
     If a show is found, the column 'insonarr' is set to true
     """
-
+    settings = Setting.load()
     if settings.SONARR_OK:
         endpoint = "/series"
         url = settings.SONARR_URL + endpoint + "?apikey=" + settings.SONARR_APIKEY
@@ -110,109 +115,102 @@ def HelperUpdateSonarr():
             ).update(insonarr=True)
 
 # checked for 0.2.0
-# TODO: Refactor:
-# On every run it should only get one page of results
-# if the current page is empty, get the last one
-# thus it can be run all 5 minutes and doesn't put too much 
-# pressure on the side.
-# we also don't need a firstrun command, since the page will fill itself uo
-@db_periodic_task(crontab(hour='*/3'))
+@background(schedule=180)
 def HelperUpdateTVMaze():
     """
     TVMazes update API provides tv shows in paged manner,
     every page contains 250 shows, leaving spaces if shows are deleted.
     the updateTvMaze function catches up from last run and gets the new shows.
     """
+    settings = Setting.load()
+    statuscode = 0
     page = settings.page
     if not page:
         page = 0
-    statuscode = 200
-    while statuscode == 200:  # as long as results are delivered
+    while statuscode != 200:
         url = "http://api.tvmaze.com/shows?page=" + str(page)
         statuscode, shows = _requestURL(url)
-        lstGenres = list()
-        for show in shows:
-            if show['language'] is not None:
-                dbLanguage, _ = Language.objects.get_or_create(language=show['language'])
-            else:
-                dbLanguage = None
-            for genre in show["genres"]:
-                dbGenre, _ = Genre.objects.get_or_create(genre=genre)
-                lstGenres.append(dbGenre)
-            if show['type'] is not None:
-                dbType, _ = ShowType.objects.get_or_create(type=show['type'])
-            else:
-                dbType = None
-            dbStatus, _ = Status.objects.get_or_create(status=show['status'])
-            if show['network'] is not None:
-                dbCountry, _ = Country.objects.get_or_create(
-                    country=show['network']['country']['name'],
-                    code=show['network']['country']['code'],
-                    timezone=show['network']['country']['timezone']
-                )
-                dbNetwork, _ = Network.objects.get_or_create(
-                    tvmaze_id=show['network']['id'],
-                    network=show['network']['name'],
-                    country=dbCountry
-                )
-            else:
-                dbNetwork = None
-            if show['webChannel'] is not None:
-                if show['webChannel']['country'] is not None:
-                    dbCountry, _ = Country.objects.get_or_create(
-                        country=show['webChannel']['country']['name'],
-                        code=show['webChannel']['country']['code'],
-                        timezone=show['webChannel']['country']['timezone']
-                    )
-                else:
-                    dbCountry = None
-                dbWebchannel, _ = Webchannel.objects.get_or_create(
-                    tvmaze_id=show['webChannel']['id'],
-                    name=show['webChannel']['name'],
-                    country=dbCountry
-                )
-            else:
-                dbWebchannel = None
-            if show['runtime'] is not None:
-                runtime = int(show['runtime'])
-            else:
-                runtime = 0
-            if show['premiered'] is not None:
-                premiere_obj = datetime.datetime.strptime(show['premiered'], '%Y-%m-%d')
-                premiere = make_aware(premiere_obj)  # make timestamp timezone aware
-            else:
-                premiere = None
-            dbShow, created = Show.objects.get_or_create(
-                tvmaze_id=show['id'],
-                defaults={
-                    'url': show['url'],
-                    'name': show['name'],
-                    'type': dbType,
-                    'language': dbLanguage,
-                    'status': dbStatus,
-                    'runtime': runtime,
-                    'premiere': premiere,
-                    'network': dbNetwork,
-                    'webchannel': dbWebchannel,
-                    'tvrage_id': show['externals']['tvrage'],
-                    'thetvdb_id': show['externals']['thetvdb'],
-                    'imdb_id': show['externals']['imdb']
-                }
+        page -= 1
+    lstGenres = list()
+    for show in shows:
+        if show['language'] is not None:
+            dbLanguage, _ = Language.objects.get_or_create(language=show['language'])
+        else:
+            dbLanguage = None
+        for genre in show["genres"]:
+            dbGenre, _ = Genre.objects.get_or_create(genre=genre)
+            lstGenres.append(dbGenre)
+        if show['type'] is not None:
+            dbType, _ = ShowType.objects.get_or_create(type=show['type'])
+        else:
+            dbType = None
+        dbStatus, _ = Status.objects.get_or_create(status=show['status'])
+        if show['network'] is not None:
+            dbCountry, _ = Country.objects.get_or_create(
+                country=show['network']['country']['name'],
+                code=show['network']['country']['code'],
+                timezone=show['network']['country']['timezone']
             )
-            if created:
-                logger.info("New show added: {}".format(show['name']))
+            dbNetwork, _ = Network.objects.get_or_create(
+                tvmaze_id=show['network']['id'],
+                network=show['network']['name'],
+                country=dbCountry
+            )
+        else:
+            dbNetwork = None
+        if show['webChannel'] is not None:
+            if show['webChannel']['country'] is not None:
+                dbCountry, _ = Country.objects.get_or_create(
+                    country=show['webChannel']['country']['name'],
+                    code=show['webChannel']['country']['code'],
+                    timezone=show['webChannel']['country']['timezone']
+                )
             else:
-                logger.info("Show already in DB: {}".format(show['name']))
-            for lstGenre in lstGenres:
-                lstGenre.shows.add(dbShow)
-            dbShow.save()
-            lstGenres.clear()
-        page += 1
-        
-        settings.page = page
-        settings.save()
-    page -= 2
+                dbCountry = None
+            dbWebchannel, _ = Webchannel.objects.get_or_create(
+                tvmaze_id=show['webChannel']['id'],
+                name=show['webChannel']['name'],
+                country=dbCountry
+            )
+        else:
+            dbWebchannel = None
+        if show['runtime'] is not None:
+            runtime = int(show['runtime'])
+        else:
+            runtime = 0
+        if show['premiered'] is not None:
+            premiere_obj = datetime.datetime.strptime(show['premiered'], '%Y-%m-%d')
+            premiere = make_aware(premiere_obj)  # make timestamp timezone aware
+        else:
+            premiere = None
+        dbShow, created = Show.objects.get_or_create(
+            tvmaze_id=show['id'],
+            defaults={
+                'url': show['url'],
+                'name': show['name'],
+                'type': dbType,
+                'language': dbLanguage,
+                'status': dbStatus,
+                'runtime': runtime,
+                'premiere': premiere,
+                'network': dbNetwork,
+                'webchannel': dbWebchannel,
+                'tvrage_id': show['externals']['tvrage'],
+                'thetvdb_id': show['externals']['thetvdb'],
+                'imdb_id': show['externals']['imdb']
+            }
+        )
+        if created:
+            logger.info("New show added: {}".format(show['name']))
+        else:
+            logger.info("Show already in DB: {}".format(show['name']))
+        for lstGenre in lstGenres:
+            lstGenre.shows.add(dbShow)
+        dbShow.save()
+        lstGenres.clear()
+    page += 1
     settings.page = page
+    logger.debug(f"Settings: {settings}")
     settings.save()
 
 #checked for 0.2.0
@@ -289,8 +287,9 @@ def updateSingleShow(tvmaze_id):
         logger.info("Show '{}' updated.".format(show['name']))
 
 #checked for 0.2.0
-@db_periodic_task(crontab(hour='*/24'))
+@background(schedule=10800)
 def HelperUpdateShows():
+    settings = Setting.load()
     url = "http://api.tvmaze.com/updates/shows"
     r = requests.get(url)
     if r.status_code == 200:
@@ -312,14 +311,17 @@ def HelperUpdateShows():
 #checked for 0.2.0
 # Is it really necessary to check all 5 Minutes?
 # TODO: Only get new profiles, if settings page is opened
-@db_periodic_task(crontab(minute='*/5'))
+@background(schedule=360))
 def helperGetSonarrProfiles():
+    settings = Setting.load()
     endpoint = "/profile/"
     url = settings.SONARR_URL + endpoint + "?apikey=" + settings.SONARR_APIKEY
     statuscode, sonarr = _requestURL(url)
-    Profile.objects.all().delete()  # First delete all current profiles
-    for s in sonarr:
-        dbProfile, _ = Profile.objects.get_or_create(
-            profile=s['name'],
-            profile_id=s['id']
-        )
+    if statuscode == 200:
+        Profile.objects.all().delete()  # First delete all current profiles
+        
+        for s in sonarr:
+            dbProfile, _ = Profile.objects.get_or_create(
+                profile=s['name'],
+                profile_id=s['id']
+            )
